requirements.txt
Flask==2.3.3
mysql-connector-python==8.2.0
PyMySQL==1.1.0
qrcode==7.4.2
Pillow==10.0.1
python-dotenv==1.0.0
pandas==2.1.1
openpyxl==3.1.2
apscheduler==3.9.1

